# order-service/src/main/java/com/itq_group/order_service/dto/OrderInfoDto.java
package com.itq_group.order_service.dto;

import com.fasterxml.jackson.annotation.JsonInclude;
import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
@JsonInclude(JsonInclude.Include.NON_NULL)
public class OrderInfoDto {
    private Long id;
}